Graphics.X11.Xft

Documentation

data XftColor

xftcolor_pixel :: XftColor -> IO Int

allocaXftColor :: (Ptr XftColor -> IO a) -> IO a

withXftColorName :: Display -> Visual -> Colormap -> String -> (XftColor -> IO a) -> IO a

withXftColorValue :: Display -> Visual -> Colormap -> XRenderColor -> (XftColor -> IO a) -> IO a

data XftDraw

withXftDraw :: Display -> Drawable -> Visual -> Colormap -> (XftDraw -> IO a) -> IO a

xftDrawCreate :: Display -> Drawable -> Visual -> Colormap -> IO XftDraw

xftDrawCreateBitmap :: Display -> Pixmap -> IO XftDraw

xftDrawCreateAlpha :: Integral a => Display -> Pixmap -> a -> IO XftDraw

xftDrawChange :: XftDraw -> Drawable -> IO ()

xftDrawDisplay :: XftDraw -> IO Display

xftDrawDrawable :: XftDraw -> IO Drawable

xftDrawColormap :: XftDraw -> IO Colormap

xftDrawVisual :: XftDraw -> IO Visual

xftDrawDestroy :: XftDraw -> IO ()

data XftFont

xftfont_ascent :: XftFont -> IO Int

xftfont_max_ascent :: NonEmpty XftFont -> IO Int

xftfont_descent :: XftFont -> IO Int

xftfont_max_descent :: NonEmpty XftFont -> IO Int

xftfont_height :: XftFont -> IO Int

xftfont_max_height :: NonEmpty XftFont -> IO Int

xftfont_max_advance_width :: XftFont -> IO Int

xftFontOpen :: Display -> Screen -> String -> IO XftFont

xftFontOpenXlfd :: Display -> Screen -> String -> IO XftFont

xftLockFace :: XftFont -> IO ()

xftUnlockFace :: XftFont -> IO ()

xftFontCopy :: Display -> XftFont -> IO XftFont

xftFontClose :: Display -> XftFont -> IO ()

xftDrawGlyphs :: (Integral a, Integral b, Integral c) => XftDraw -> XftColor -> XftFont -> b -> c -> [a] -> IO ()

xftDrawString :: (Integral a, Integral b) => XftDraw -> XftColor -> XftFont -> a -> b -> String -> IO ()

xftDrawStringFallback :: XftDraw -> XftColor -> [XftFont] -> Int -> Int -> String -> IO ()

Like xftDrawString, but fall back to another font in the given list if necessary (i.e., should a character not be drawable with the currently selected font).

xftTextExtents :: Display -> XftFont -> String -> IO XGlyphInfo

xftTextAccumExtents :: Display -> [XftFont] -> String -> IO XGlyphInfo

Like xftTextExtents but for multiple fonts. Return accumulative extents using appropriate fonts for each part of string.

xftDrawRect :: (Integral a, Integral b, Integral c, Integral d) => XftDraw -> XftColor -> a -> b -> c -> d -> IO ()

xftDrawSetClipRectangles :: XftDraw -> Int -> Int -> [Rectangle] -> IO Bool

xftDrawSetSubwindowMode :: Integral a => XftDraw -> a -> IO ()

xftInitFtLibrary :: IO ()